Render LEDs using a point-spread function
The point spread function models the lens flare and blur of the human
visual system. The model is taken directly from [1].

In an ideal world, every pixel would take every LED into account during
rendering. However, this involves 1K PSF calculations per pixel for a
32x32 grid of LEDs, which is not feasible on e.g. a 13" MacBook Pro.

In order to achieve reasonable performance, each pixel only evaluates
PSFs for its 49 nearest neighbors. This provides good visual fidelity at
a fraction of the cost of the complete approach.

The pixel shader is also used to render LEDs that are not turned on.
These LEDs are represented as filled gray circles.